moved coro_utils functions to  scoob_utils for dedicated scoob hardware control
