+++ lina/experimental/aefc_2dm.py
@@ -1,5 +1,5 @@
 from lina.math_module import xp, xcipy, ensure_np_array
-from lina import utils, coro_utils, pwp
+from lina import utils, pwp
 
 import numpy as np
 import scipy
@@ -118,7 +118,7 @@ def run(
 
         metric_im = take_im_fun(**take_im_params)
         metric_im_ni = metric_im if normalize_metric_fun is None else normalize_metric_fun(metric_im, **normalize_metric_params)
-        contrast = coro_utils.compute_contrast(metric_im_ni, wfs_mask)
+        contrast = utils.compute_contrast(metric_im_ni, wfs_mask)
 
         aefc_data['raw_images'].append(copy.copy(metric_im))
         aefc_data['ni_images'].append(copy.copy(metric_im_ni))
@@ -247,7 +247,7 @@ def run_mw(
 
         metric_im = take_im_fun(**take_im_params)
         metric_im_ni = metric_im if normalize_metric_fun is None else normalize_metric_fun(metric_im, **normalize_metric_params)
-        contrast = coro_utils.compute_contrast(metric_im_ni, wfs_mask)
+        contrast = utils.compute_contrast(metric_im_ni, wfs_mask)
 
         aefc_data['raw_images'].append(copy.copy(metric_im))
         aefc_data['ni_images'].append(copy.copy(metric_im_ni))
